@@ -9,22 +9,25 @@ import * as $cluster from "cluster"
 import * as $path from "path";
 import * as $fs from "fs";
 import * as ObjectUtils from "jumbo-core/utils/object";
-
-// Default config values
-let defaultConfig = require("jumbo-core/default-config.js");
+import {dirname} from "jumbo-core/utils/path";
 
 // Start timer measuring application load time
-if ($cluster.isMaster) {
+if ($cluster.isMaster)
+{
 	console.time("Application Master load-time: ");
-} else {
+}
+else
+{
 	console.time("Application Worker " + $cluster.worker.id + " load-time: ");
 }
+
+const DIRNAME = dirname(module);
 
 /**
  * Base project directory
  * @type {string}
  */
-const PROJECT_DIR = $path.dirname(require.main.filename).toLowerCase();
+const PROJECT_DIR = $path.dirname(require.main.filename);//.toLowerCase();
 
 /**
  * Number of milisecond in day
@@ -33,45 +36,16 @@ const PROJECT_DIR = $path.dirname(require.main.filename).toLowerCase();
 const DAY_MS = 24 * 60 * 60 * 1000;
 
 /**
- * List of folders which should Jumbo application contains
- */
-const FRAMEWORK_FOLDER_STRUCTURE = [
-	$path.join(PROJECT_DIR, "app"),
-	$path.join(PROJECT_DIR, "app", "controllers"),
-	$path.join(PROJECT_DIR, "app", "sub-apps"),
-	$path.join(PROJECT_DIR, "app", "services"),
-	$path.join(PROJECT_DIR, "app", "models"),
-	$path.join(PROJECT_DIR, "app", "templates"),
-	$path.join(PROJECT_DIR, "app", "tests"),
-
-	$path.join(PROJECT_DIR, "data"),
-	$path.join(PROJECT_DIR, "data", "uploads"),
-	$path.join(PROJECT_DIR, "data", "logs"),
-	$path.join(PROJECT_DIR, "data", "errors"),
-
-	$path.join(PROJECT_DIR, "public"),
-	// $path.join(base, "public", "styles"),
-	// $path.join(base, "public", "scripts"),
-	// $path.join(base, "public", "images"),
-
-	$path.join(PROJECT_DIR, "temp"),
-	$path.join(PROJECT_DIR, "temp", "cache"),
-	$path.join(PROJECT_DIR, "temp", "sessions")
-];
-
-/**
  * Declaration of global object/namespace Jumbo which represents
  * base framework object for accessing framework data
- * @namespace Jumbo
- * @global
  */
-const Jumbo: JumboNamespace = {
-	/** @type {defaultConfig} */
+const JumboGlobalNamespace = {
+	/** @type {ApplicationConfig} */
 	config: {},
 	CONFIG_PATH: $path.join(PROJECT_DIR, "config.js"),
-	CFG_PATH: $path.join(__dirname, "config.js").toLowerCase(),
+	CFG_PATH: $path.join(DIRNAME, "config.js"),
 	BASE_DIR: PROJECT_DIR,
-	CORE_DIR: __dirname.toLowerCase(),
+	CORE_DIR: DIRNAME,
 	PUBLIC_DIR: $path.join(PROJECT_DIR, "public"),
 	APP_DIR: $path.join(PROJECT_DIR, "app"),
 	SUB_APP_DIR: $path.join(PROJECT_DIR, "app", "sub-apps"),
@@ -80,23 +54,12 @@ const Jumbo: JumboNamespace = {
 	UPLOAD_DIR: $path.resolve(PROJECT_DIR, "data", "uploads"),
 	CACHE_DIR: $path.resolve(PROJECT_DIR, "temp", "cache"),
 	SESSION_DIR: $path.resolve(PROJECT_DIR, "temp", "sessions"),
-
-	Application: {},
-	Adapters: {},
-	Autoloader: {},
-	Base: {},
-	Cluster: {},
-	Exceptions: {},
-	Ioc: {},
-	Logging: {},
-	Sync: {},
-	Utils: {},
-	Validation: {},
 };
 
-(<any>global).Jumbo = Jumbo;
+(<any>global).Jumbo = JumboGlobalNamespace;
 
-class Loader {
+class Loader
+{
 	//region Fields
 
 	/**
@@ -107,7 +70,7 @@ class Loader {
 
 	/**
 	 * Config; Filled from checkConfig()
-	 * @type {defaultConfig}
+	 * @type {ApplicationConfig}
 	 */
 	private config: any;
 
@@ -118,11 +81,13 @@ class Loader {
 	/**
 	 * Initialize application
 	 */
-	public static initializeApplication() {
+	public static initializeApplication()
+	{
 		let loader = new Loader();
 		loader.initialize();
 
-		if ($cluster.isMaster) {
+		if ($cluster.isMaster)
+		{
 			loader.deleteCachedFiles();
 			loader.deleteOldSessions();
 		}
@@ -160,13 +125,16 @@ class Loader {
 	/**
 	 * Delete cached filed
 	 */
-	public deleteCachedFiles() {
-		$fs.readdir(Jumbo.CACHE_DIR, (err, files) => {
+	public deleteCachedFiles()
+	{
+		$fs.readdir(JumboGlobalNamespace.CACHE_DIR, (err, files) => {
 			let i = 0;
 
-			for (let fileName of files) {
-				if (fileName.slice(-9) == ".tplcache") {
-					let file = $path.join(Jumbo.CACHE_DIR, fileName);
+			for (let fileName of files)
+			{
+				if (fileName.slice(-9) == ".tplcache")
+				{
+					let file = $path.join(JumboGlobalNamespace.CACHE_DIR, fileName);
 
 					// remove file
 					$fs.unlink(file, () => { });
@@ -182,16 +150,20 @@ class Loader {
 	/**
 	 * Delete old sessions
 	 */
-	public deleteOldSessions() {
-		$fs.readdir(Jumbo.SESSION_DIR, (err, files) => {
+	public deleteOldSessions()
+	{
+		$fs.readdir(JumboGlobalNamespace.SESSION_DIR, (err, files) => {
 			let sessionLimitTime = (new Date().getTime() - Jumbo.config.session.sessionLifetime * DAY_MS);
 
-			for (let fileName of files) {
-				if (fileName.slice(-8) == ".session") {
-					let file = $path.join(Jumbo.SESSION_DIR, fileName);
+			for (let fileName of files)
+			{
+				if (fileName.slice(-8) == ".session")
+				{
+					let file = $path.join(JumboGlobalNamespace.SESSION_DIR, fileName);
 					let stats = $fs.statSync(file);
 
-					if (stats.birthtime.getTime() < sessionLimitTime) {
+					if (stats.birthtime.getTime() < sessionLimitTime)
+					{
 						// remove session file
 						// noinspection JSUnusedLocalSymbols
 						$fs.unlink(file, (err) => {});
@@ -206,7 +178,8 @@ class Loader {
 	/**
 	 * Initialize autoloader
 	 */
-	public initAutoloader() {
+	public initAutoloader()
+	{
 		// Load the Autoloader
 		const autoloader = require("jumbo-core/autoloader/autoloader");
 
@@ -219,19 +192,22 @@ class Loader {
 		// Add items from core directory to namespace/object Jumbo
 		let objs = Object.getOwnPropertyNames(autoloader.Core);
 		let c = objs.length;
-		for (let p = 0; p < c; p++) {
-			Jumbo[objs[p]] = autoloader.Core[objs[p]];
+		for (let p = 0; p < c; p++)
+		{
+			JumboGlobalNamespace[objs[p]] = autoloader.Core[objs[p]];
 		}
 	}
 
 	/**
 	 * Provede inicializaci
 	 */
-	public initialize() {
+	public initialize()
+	{
 		this.checkConfig();
 		this.checkAppStructure();
 
-		if (this.exitStatus) {
+		if (this.exitStatus)
+		{
 			process.exit(0);
 		}
 	}
@@ -243,33 +219,66 @@ class Loader {
 	/**
 	 * Check that all framework directories exists
 	 */
-	private checkAppStructure() {
-		FRAMEWORK_FOLDER_STRUCTURE.forEach(function (p) {
-			try {
+	private checkAppStructure()
+	{
+		[
+			$path.join(PROJECT_DIR, "app"),
+			$path.join(PROJECT_DIR, "app", "controllers"),
+			$path.join(PROJECT_DIR, "app", "sub-apps"),
+			$path.join(PROJECT_DIR, "app", "services"),
+			$path.join(PROJECT_DIR, "app", "models"),
+			$path.join(PROJECT_DIR, "app", "templates"),
+			// $path.join(PROJECT_DIR, "app", "tests"),
+
+			$path.join(PROJECT_DIR, "data"),
+			$path.join(PROJECT_DIR, "data", "uploads"),
+			$path.join(PROJECT_DIR, "data", "logs"),
+			$path.join(PROJECT_DIR, "data", "errors"),
+
+			$path.join(PROJECT_DIR, "public"),
+			// $path.join(base, "public", "styles"),
+			// $path.join(base, "public", "scripts"),
+			// $path.join(base, "public", "images"),
+
+			$path.join(PROJECT_DIR, "temp"),
+			$path.join(PROJECT_DIR, "temp", "cache"),
+			$path.join(PROJECT_DIR, "temp", "sessions")
+		].forEach(function (p) {
+			try
+			{
 				let stat = $fs.lstatSync(p);
 
-				if (!stat.isDirectory()) {
+				if (!stat.isDirectory())
+				{
 					console.error(`[ERROR] Structure directory '${p}' not found.`);
 					this.exitStatus = true;
 				}
-			} catch (ex) {
+			}
+			catch (ex)
+			{
 				this.exitStatus = true;
 			}
 		});
 	}
 
-	private isInConfig(section, ...properties) {
+	private isInConfig(section, ...properties)
+	{
 		let sect = this.config[section];
 		let succ = true;
 
-		if (sect === undefined) {
+		if (sect === undefined)
+		{
 			console.error(`[ERROR] Config file is corrupted. Section '${section}' is missing.`);
 			succ = false;
-		} else {
-			for (let prop in properties) {
+		}
+		else
+		{
+			for (let prop in properties)
+			{
 				prop = properties[prop];
 
-				if (!sect.hasOwnProperty(prop)) {
+				if (!sect.hasOwnProperty(prop))
+				{
 					console.error(`[ERROR] Config file is corrupted. Property '${prop}' is missing in section '${section}'.`);
 					succ = false;
 				}
@@ -282,23 +291,27 @@ class Loader {
 	/**
 	 * Load config, check it and prepare object with readonly properties and put it into global Jumbo
 	 */
-	private checkConfigSections() {
+	private checkConfigSections()
+	{
 		if (
 			this.isInConfig("protocol", "protocol", "privateKey", "certificate", "pfx", "passphrase")
 			&& this.isInConfig("clustering", "numberOfWorkers")
 			&& this.isInConfig("cache", "enabled", "memoryCacheSizeLimit")
 			&& this.isInConfig("session", "sessionsCookieName", "sessionLifetime", "memorySizeLimit", "justInMemory")
 			&& this.isInConfig("log", "enabled", "level")
-			&& this.isInConfig("doTestsAfterRun")
+			// && this.isInConfig("doTestsAfterRun")
 			&& this.isInConfig("maxRequestPerSecond")
 			&& this.isInConfig("maxPostDataSize")
 			&& this.isInConfig("deployment")
 			&& this.isInConfig("debugMode")
 			&& this.isInConfig("DOSPrevention", "enabled", "blockTime", "maxRequestPerIP")
 			&& this.isInConfig("globalization", "enabled")
-		) {
-			Jumbo.config = ObjectUtils.freeze(this.config, 2);
-		} else {
+		)
+		{
+			JumboGlobalNamespace.config = ObjectUtils.freeze(this.config, 2);
+		}
+		else
+		{
 			// ...
 		}
 	}
@@ -306,18 +319,26 @@ class Loader {
 	/**
 	 * Check that config exists
 	 */
-	private checkConfig() {
-		let confPath = $path.join(PROJECT_DIR, "config.js");
-
-		if (!$fs.lstatSync(confPath).isFile()) {
+	private checkConfig()
+	{
+		if (!$fs.lstatSync(Jumbo.CONFIG_PATH).isFile())
+		{
 			this.exitStatus = true;
-			console.error("Application config '" + confPath + "' not found.");
+			console.error("Application config '" + Jumbo.CONFIG_PATH + "' not found.");
 		}
 
-		try {
-			this.config = ObjectUtils.assign(defaultConfig, require(confPath));
+		try
+		{
+			// Default config values
+			let defaultConfig = require("jumbo-core/default-config.js");
+
+			// Extend default config with app config
+			this.config = ObjectUtils.assign(defaultConfig, require(Jumbo.CONFIG_PATH));
+
 			this.checkConfigSections();
-		} catch (ex) {
+		}
+		catch (ex)
+		{
 			this.exitStatus = true;
 			console.error("Config JSON invalid.", ex);
 		}
@@ -326,13 +347,17 @@ class Loader {
 	//endregion
 }
 
-console.log("*******************************");
-console.log("**");
-console.log("** JumboJS, booting up...");
-console.log("**");
-console.log("*******************************");
+if ($cluster.isMaster)
+{
+	console.log("*******************************");
+	console.log("**");
+	console.log("** JumboJS, booting up...");
+	console.log("**");
+	console.log("*******************************");
+}
 
 Loader.initializeApplication();
 
 import {Application} from "./application/Application";
+
 export const application: Application = (<any>global).Application;
