@@ -4,38 +4,22 @@ const $cluster = require("cluster");
 const $path = require("path");
 const $fs = require("fs");
 const ObjectUtils = require("jumbo-core/utils/object");
-let defaultConfig = require("jumbo-core/default-config.js");
+const path_1 = require("jumbo-core/utils/path");
 if ($cluster.isMaster) {
     console.time("Application Master load-time: ");
 }
 else {
     console.time("Application Worker " + $cluster.worker.id + " load-time: ");
 }
-const PROJECT_DIR = $path.dirname(require.main.filename).toLowerCase();
+const DIRNAME = path_1.dirname(module);
+const PROJECT_DIR = $path.dirname(require.main.filename);
 const DAY_MS = 24 * 60 * 60 * 1000;
-const FRAMEWORK_FOLDER_STRUCTURE = [
-    $path.join(PROJECT_DIR, "app"),
-    $path.join(PROJECT_DIR, "app", "controllers"),
-    $path.join(PROJECT_DIR, "app", "sub-apps"),
-    $path.join(PROJECT_DIR, "app", "services"),
-    $path.join(PROJECT_DIR, "app", "models"),
-    $path.join(PROJECT_DIR, "app", "templates"),
-    $path.join(PROJECT_DIR, "app", "tests"),
-    $path.join(PROJECT_DIR, "data"),
-    $path.join(PROJECT_DIR, "data", "uploads"),
-    $path.join(PROJECT_DIR, "data", "logs"),
-    $path.join(PROJECT_DIR, "data", "errors"),
-    $path.join(PROJECT_DIR, "public"),
-    $path.join(PROJECT_DIR, "temp"),
-    $path.join(PROJECT_DIR, "temp", "cache"),
-    $path.join(PROJECT_DIR, "temp", "sessions")
-];
-const Jumbo = {
+const JumboGlobalNamespace = {
     config: {},
     CONFIG_PATH: $path.join(PROJECT_DIR, "config.js"),
-    CFG_PATH: $path.join(__dirname, "config.js").toLowerCase(),
+    CFG_PATH: $path.join(DIRNAME, "config.js"),
     BASE_DIR: PROJECT_DIR,
-    CORE_DIR: __dirname.toLowerCase(),
+    CORE_DIR: DIRNAME,
     PUBLIC_DIR: $path.join(PROJECT_DIR, "public"),
     APP_DIR: $path.join(PROJECT_DIR, "app"),
     SUB_APP_DIR: $path.join(PROJECT_DIR, "app", "sub-apps"),
@@ -44,19 +28,8 @@ const Jumbo = {
     UPLOAD_DIR: $path.resolve(PROJECT_DIR, "data", "uploads"),
     CACHE_DIR: $path.resolve(PROJECT_DIR, "temp", "cache"),
     SESSION_DIR: $path.resolve(PROJECT_DIR, "temp", "sessions"),
-    Application: {},
-    Adapters: {},
-    Autoloader: {},
-    Base: {},
-    Cluster: {},
-    Exceptions: {},
-    Ioc: {},
-    Logging: {},
-    Sync: {},
-    Utils: {},
-    Validation: {},
 };
-global.Jumbo = Jumbo;
+global.Jumbo = JumboGlobalNamespace;
 class Loader {
     constructor() {
         this.exitStatus = false;
@@ -77,11 +50,11 @@ class Loader {
         };
     }
     deleteCachedFiles() {
-        $fs.readdir(Jumbo.CACHE_DIR, (err, files) => {
+        $fs.readdir(JumboGlobalNamespace.CACHE_DIR, (err, files) => {
             let i = 0;
             for (let fileName of files) {
                 if (fileName.slice(-9) == ".tplcache") {
-                    let file = $path.join(Jumbo.CACHE_DIR, fileName);
+                    let file = $path.join(JumboGlobalNamespace.CACHE_DIR, fileName);
                     $fs.unlink(file, () => { });
                     i++;
                 }
@@ -90,11 +63,11 @@ class Loader {
         });
     }
     deleteOldSessions() {
-        $fs.readdir(Jumbo.SESSION_DIR, (err, files) => {
+        $fs.readdir(JumboGlobalNamespace.SESSION_DIR, (err, files) => {
             let sessionLimitTime = (new Date().getTime() - Jumbo.config.session.sessionLifetime * DAY_MS);
             for (let fileName of files) {
                 if (fileName.slice(-8) == ".session") {
-                    let file = $path.join(Jumbo.SESSION_DIR, fileName);
+                    let file = $path.join(JumboGlobalNamespace.SESSION_DIR, fileName);
                     let stats = $fs.statSync(file);
                     if (stats.birthtime.getTime() < sessionLimitTime) {
                         $fs.unlink(file, (err) => { });
@@ -110,7 +83,7 @@ class Loader {
         let objs = Object.getOwnPropertyNames(autoloader.Core);
         let c = objs.length;
         for (let p = 0; p < c; p++) {
-            Jumbo[objs[p]] = autoloader.Core[objs[p]];
+            JumboGlobalNamespace[objs[p]] = autoloader.Core[objs[p]];
         }
     }
     initialize() {
@@ -121,7 +94,22 @@ class Loader {
         }
     }
     checkAppStructure() {
-        FRAMEWORK_FOLDER_STRUCTURE.forEach(function (p) {
+        [
+            $path.join(PROJECT_DIR, "app"),
+            $path.join(PROJECT_DIR, "app", "controllers"),
+            $path.join(PROJECT_DIR, "app", "sub-apps"),
+            $path.join(PROJECT_DIR, "app", "services"),
+            $path.join(PROJECT_DIR, "app", "models"),
+            $path.join(PROJECT_DIR, "app", "templates"),
+            $path.join(PROJECT_DIR, "data"),
+            $path.join(PROJECT_DIR, "data", "uploads"),
+            $path.join(PROJECT_DIR, "data", "logs"),
+            $path.join(PROJECT_DIR, "data", "errors"),
+            $path.join(PROJECT_DIR, "public"),
+            $path.join(PROJECT_DIR, "temp"),
+            $path.join(PROJECT_DIR, "temp", "cache"),
+            $path.join(PROJECT_DIR, "temp", "sessions")
+        ].forEach(function (p) {
             try {
                 let stat = $fs.lstatSync(p);
                 if (!stat.isDirectory()) {
@@ -158,26 +146,25 @@ class Loader {
             && this.isInConfig("cache", "enabled", "memoryCacheSizeLimit")
             && this.isInConfig("session", "sessionsCookieName", "sessionLifetime", "memorySizeLimit", "justInMemory")
             && this.isInConfig("log", "enabled", "level")
-            && this.isInConfig("doTestsAfterRun")
             && this.isInConfig("maxRequestPerSecond")
             && this.isInConfig("maxPostDataSize")
             && this.isInConfig("deployment")
             && this.isInConfig("debugMode")
             && this.isInConfig("DOSPrevention", "enabled", "blockTime", "maxRequestPerIP")
             && this.isInConfig("globalization", "enabled")) {
-            Jumbo.config = ObjectUtils.freeze(this.config, 2);
+            JumboGlobalNamespace.config = ObjectUtils.freeze(this.config, 2);
         }
         else {
         }
     }
     checkConfig() {
-        let confPath = $path.join(PROJECT_DIR, "config.js");
-        if (!$fs.lstatSync(confPath).isFile()) {
+        if (!$fs.lstatSync(Jumbo.CONFIG_PATH).isFile()) {
             this.exitStatus = true;
-            console.error("Application config '" + confPath + "' not found.");
+            console.error("Application config '" + Jumbo.CONFIG_PATH + "' not found.");
         }
         try {
-            this.config = ObjectUtils.assign(defaultConfig, require(confPath));
+            let defaultConfig = require("jumbo-core/default-config.js");
+            this.config = ObjectUtils.assign(defaultConfig, require(Jumbo.CONFIG_PATH));
             this.checkConfigSections();
         }
         catch (ex) {
@@ -186,11 +173,13 @@ class Loader {
         }
     }
 }
-console.log("*******************************");
-console.log("**");
-console.log("** JumboJS, booting up...");
-console.log("**");
-console.log("*******************************");
+if ($cluster.isMaster) {
+    console.log("*******************************");
+    console.log("**");
+    console.log("** JumboJS, booting up...");
+    console.log("**");
+    console.log("*******************************");
+}
 Loader.initializeApplication();
 exports.application = global.Application;
 //# sourceMappingURL=loader.js.map
